refactor packages in core module
