+++ app/src/main/java/com/fayaz/todo_jc/main/MainActivity.kt
@@ -4,7 +4,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.viewModels
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.fayaz.todo_jc.core.logger.actions.ShowOnboardingActivity
+import com.fayaz.todo_jc.core.actions.ShowOnboardingActivity
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
