+++ core/src/main/java/com/fayaz/todo_jc/core/actions/ShowDashboardActivity.kt
@@ -1,4 +1,4 @@
-package com.fayaz.todo_jc.core.logger.actions
+package com.fayaz.todo_jc.core.actions
 
 import android.app.Activity
