+++ core/src/main/java/com/fayaz/todo_jc/core/base/StateViewModel.kt
@@ -1,12 +1,9 @@
-package com.fayaz.todo_jc.core.logger.base
+package com.fayaz.todo_jc.core.base
 
 import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
 
 abstract class StateViewModel<UiEvent: ViewEvent, UiState: ViewState>: ViewModel() {
   private val initialState: UiState by lazy { setInitialState() }
