+++ core/src/main/java/com/fayaz/todo_jc/core/base/ViewEvent.kt
@@ -1,0 +1,3 @@
+package com.fayaz.todo_jc.core.base
+
+open class ViewEvent
