--- core/src/main/java/com/fayaz/todo_jc/core/logger/base/ViewState.kt
@@ -1,3 +1,0 @@
-package com.fayaz.todo_jc.core.logger.base
-
-open class ViewState
