--- core/src/main/java/com/fayaz/todo_jc/core/logger/di/ActionsModule.kt
@@ -1,14 +1,0 @@
-package com.fayaz.todo_jc.core.logger.di
-
-import com.fayaz.todo_jc.core.logger.actions.ShowOnboardingActivity
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-//@Module
-//@InstallIn(SingletonComponent::class)
-//object ActionsModule  {
-//  @Provides
-//  fun provideInterface() = ShowOnboardingActivity
-//}
