+++ features/dashboard/src/main/java/com/fayaz/todo_jc/features/dashboard/ShowDashboardActivityImpl.kt
@@ -2,7 +2,7 @@ package com.fayaz.todo_jc.features.dashboard
 
 import android.app.Activity
 import android.content.Intent
-import com.fayaz.todo_jc.core.logger.actions.ShowDashboardActivity
+import com.fayaz.todo_jc.core.actions.ShowDashboardActivity
 import com.fayaz.todo_jc.features.dashboard.ui.DashboardActivity
 import javax.inject.Inject
