+++ features/dashboard/src/main/java/com/fayaz/todo_jc/features/dashboard/ui/DashboardEvent.kt
@@ -1,6 +1,6 @@
 package com.fayaz.todo_jc.features.dashboard.ui
 
-import com.fayaz.todo_jc.core.logger.base.ViewEvent
+import com.fayaz.todo_jc.core.base.ViewEvent
 
 sealed class DashboardEvent : ViewEvent() {
   object Submit : DashboardEvent()
