+++ features/dashboard/src/main/java/com/fayaz/todo_jc/features/dashboard/ui/DashboardState.kt
@@ -1,6 +1,6 @@
 package com.fayaz.todo_jc.features.dashboard.ui
 
-import com.fayaz.todo_jc.core.logger.base.ViewState
+import com.fayaz.todo_jc.core.base.ViewState
 
 data class DashboardState(
   val loading: Boolean = false,
