+++ features/onboarding/src/main/java/com/fayaz/todo_jc/features/onboarding/actions/ShowOnboardingActivityImpl.kt
@@ -2,7 +2,7 @@ package com.fayaz.todo_jc.features.onboarding.actions
 
 import android.app.Activity
 import android.content.Intent
-import com.fayaz.todo_jc.core.logger.actions.ShowOnboardingActivity
+import com.fayaz.todo_jc.core.actions.ShowOnboardingActivity
 import com.fayaz.todo_jc.features.onboarding.ui.OnboardingActivity
 import javax.inject.Inject
