+++ features/onboarding/src/main/java/com/fayaz/todo_jc/features/onboarding/di/OnboardingActivityModule.kt
@@ -1,11 +1,10 @@
 package com.fayaz.todo_jc.features.onboarding.di
 
-import com.fayaz.todo_jc.core.logger.actions.ShowOnboardingActivity
+import com.fayaz.todo_jc.core.actions.ShowOnboardingActivity
 import com.fayaz.todo_jc.features.onboarding.actions.ShowOnboardingActivityImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
-import dagger.hilt.android.scopes.ActivityScoped
 import dagger.hilt.components.SingletonComponent
 
 @Module
