+++ features/onboarding/src/main/java/com/fayaz/todo_jc/features/onboarding/ui/OnboardingActivity.kt
@@ -19,7 +19,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.fayaz.todo_jc.core.logger.actions.ShowDashboardActivity
+import com.fayaz.todo_jc.core.actions.ShowDashboardActivity
 import com.fayaz.todo_jc.design_kit.theme.AppTheme
 import com.fayaz.todo_jc.features.onboarding.R
 import com.fayaz.todo_jc.features.onboarding.data.onboardingData
